Merkle trie visitor refactoring
